Refactor: move entity to one package
