+++ src/test/java/xyz/npgw/test/common/entity/BusinessUnit.java
@@ -1,4 +1,4 @@
-package xyz.npgw.test.common.util;
+package xyz.npgw.test.common.entity;
 
 public record BusinessUnit(String merchantId, String merchantName) {
