--- src/test/java/xyz/npgw/test/common/util/Merchant.java
@@ -1,9 +1,0 @@
-package xyz.npgw.test.common.util;
-
-public record Merchant(
-        String merchantName,
-        boolean usd,
-        boolean eur,
-        boolean active
-) {
-}
