+++ src/test/java/xyz/npgw/test/run/AddAcquirerDialogTest.java
@@ -19,9 +19,9 @@ import java.util.List;
 
 import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
 import static org.testng.Assert.assertEquals;
-import static xyz.npgw.test.common.TestUtils.createAcquirer;
-import static xyz.npgw.test.common.TestUtils.deleteAcquirer;
-import static xyz.npgw.test.common.TestUtils.getAcquirer;
+import static xyz.npgw.test.common.util.TestUtils.createAcquirer;
+import static xyz.npgw.test.common.util.TestUtils.deleteAcquirer;
+import static xyz.npgw.test.common.util.TestUtils.getAcquirer;
 
 public class AddAcquirerDialogTest extends BaseTest {
