+++ src/test/java/xyz/npgw/test/run/EditBusinessUnitDialogTest.java
@@ -8,8 +8,8 @@ import io.qameta.allure.Feature;
 import io.qameta.allure.TmsLink;
 import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
-import xyz.npgw.test.common.TestUtils;
 import xyz.npgw.test.common.base.BaseTest;
+import xyz.npgw.test.common.util.TestUtils;
 import xyz.npgw.test.page.DashboardPage;
 
 import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
