+++ src/test/java/xyz/npgw/test/run/TeamPageTest.java
@@ -9,11 +9,11 @@ import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import xyz.npgw.test.common.Constants;
 import xyz.npgw.test.common.ProjectProperties;
-import xyz.npgw.test.common.TestUtils;
 import xyz.npgw.test.common.UserRole;
 import xyz.npgw.test.common.base.BaseTest;
+import xyz.npgw.test.common.entity.User;
 import xyz.npgw.test.common.provider.TestDataProvider;
-import xyz.npgw.test.common.util.User;
+import xyz.npgw.test.common.util.TestUtils;
 import xyz.npgw.test.page.DashboardPage;
 import xyz.npgw.test.page.dialog.user.AddUserDialog;
 import xyz.npgw.test.page.dialog.user.EditUserDialog;
